First
For some reason I broke this back in July.  I was trying to pull the
sample generating process out of the tone, and it required a bunch of
passing data that I wasn't prepared for.  So, I think I'm going to try
to put it back the way it was and then re-evaluate where I want to take
it.