Back to square one
Put it back the way it was, except I refined it so that the track length
is only as long as one cycle of the waveform.  Next steps are to add new
waveforms, variable pitch/freq/volume, and abstract it away.  Also,
build up storage so that we don't always have to generate them.