Refactoring Tone into Waveform + Tone + WaveformView
Intent is to be able to generate waveforms independent of playing them.
So, basically we can generate a waveform based on some input, then load
up the tone (which is effectively the playing of the waveform for some
duration), and/or pass the generated Waveform to WaveformView for
display separately.  Also added support for different shapes, like
saw/square/sine/etc.  Should probably add a list to test these out and
see how they sound.  It's also probably time to change the slider bar
layout to something more precise...  Maybe even start a dictionary of
real notes to pull frequencies from.  Also need to re-implement volume
and pitch.  FINALLY, also need to implement variable volume and pitch
during playback, which is probably easier said than done.