Add FFT + toggle harmonics for testing
Not using FFT, yet.  It still needs refactoring and optimization (ie.
can I get rid of the imaginary part since there is none in the audio).
Made harmonics -4 to 4 togglable for testing.  Added checks for
inaudible frequencies, so they don't get added in.  It's still a pain to
use this, because the waveform is only generated by selecting notes.
All other actions just manipulate parameters.  Long term it should be
fine, since all the parameters will be calculated in code and then
generated.