Waveform refactor done
There are some minor things, maybe some efficiency things I'd like to
address one day.  But, the interface is what it should be, and I think
it's time to spend some time doing something else.  Possible
improvements: Remove set_waveform function, generate the fundamental as
a single cycle, and then generate the output waveform as the longest of
the fundamental and the harmonics.  Right now the fundamental is being
generated to fit the longest.  It might be introducing some
clipping/popping noise that sucks.